+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- `--timeout` option now affects both dialing to server and RPC call (before that dialing had hard-coded 1s timeout)
+
 ## 1.0.0 - 2017-12-13
 
 ### Added